requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environ==1.0
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
simplejson==3.19.2
sqlparse
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
